Refactored the EntityBuilders
