+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/datamodel/resource/storage/simple/SimpleStorageResourceValidator.java
@@ -24,10 +24,8 @@ package net.ivoa.calycopis.datamodel.resource.storage.simple;
 
 import lombok.extern.slf4j.Slf4j;
 import net.ivoa.calycopis.datamodel.offerset.OfferSetRequestParserContext;
-import net.ivoa.calycopis.datamodel.resource.storage.AbstractStorageResourceEntity;
 import net.ivoa.calycopis.datamodel.resource.storage.AbstractStorageResourceValidator;
 import net.ivoa.calycopis.datamodel.session.ExecutionSessionEntity;
-import net.ivoa.calycopis.functional.builder.Builder;
 import net.ivoa.calycopis.functional.validator.Validator;
 import net.ivoa.calycopis.functional.validator.ValidatorTools;
 import net.ivoa.calycopis.openapi.model.IvoaAbstractStorageResource;
@@ -95,22 +93,18 @@ implements AbstractStorageResourceValidator
         IvoaSimpleStorageResource validated = new IvoaSimpleStorageResource();
 
         //
-        // Check for maximum limits,
-        // Validate the lifetime values,
+        // Check maximum size limit,
+        // Validate the lifetime syntax,
         //
 
         validated.setName(requested.getName());
         
         //
         // Everything is good.
-        // Create our result and add it to our state.
-        // TODO Need to add a reference to the builder.
+        // Create our result and add it to our context.
         if (success)
             {
-            log.debug("Success");
-
-            log.debug("Creating Builder.");
-            Builder<AbstractStorageResourceEntity> builder = new Builder<AbstractStorageResourceEntity>()
+            EntityBuilder builder = new EntityBuilder()
                 {
                 @Override
                 public SimpleStorageResourceEntity build(final ExecutionSessionEntity executionSession)
@@ -122,14 +116,12 @@ implements AbstractStorageResourceValidator
                     }
                 }; 
             
-            log.debug("Creating Result.");
             AbstractStorageResourceValidator.Result storageResult = new AbstractStorageResourceValidator.ResultBean(
                 Validator.ResultEnum.ACCEPTED,
                 validated,
                 builder
                 );
-            //
-            // Save the DataResource in the state.
+
             context.addStorageValidatorResult(
                 storageResult 
                 );
