+++ notes/zrq/20250421-01-exploring.txt
@@ -326,7 +326,7 @@ The download step needs access to the storage
 
             PodmanStorageResourceEntityFactory
             extends
-            StorageResourceEntityFactory
+            SimpleStorageResourceEntityFactory
 
             create(...)
              => PodmanStorageResourceEntity
